requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
amqp==5.1.1
aniso8601==9.0.1
apispec==5.1.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==6.2
Deprecated==1.2.13
Flask==2.0.2
flask-apispec==0.11.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.3.3
kombu==5.2.4
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
mistune==2.0.1
multidict==6.0.2
packaging==21.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
pyrsistent==0.18.0
python-dotenv==0.19.2
python-editor==1.0.4
pytz==2021.3
PyYAML==6.0
redis==4.2.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webargs==8.0.1
Werkzeug==2.0.2
wrapt==1.14.0
yarl==1.7.2
